requirements.txt
google-genai>=1.0.0
feedparser>=6.0
beautifulsoup4>=4.12
lxml>=4.9
requests>=2.31
Jinja2>=3.1
PyYAML>=6.0
python-dateutil>=2.8
